backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==1.4.54
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
bcrypt==4.0.1

apscheduler==3.11.2

streamlit==1.40.2
langchain==0.3.13
langchain-openai==0.2.14
langchain-chroma==0.1.4
langchain-community==0.3.13
chromadb==0.5.23
pypdf==5.1.0
python-decouple==3.8

numpy==1.24.3
xgboost==2.0.3
scikit-learn==1.3.2